+++ core/src/main/kotlin/sift/core/asm/signature/TypeSignature.kt
@@ -1,7 +1,7 @@
 package sift.core.asm.signature
 
-import org.objectweb.asm.Type
 import sift.core.asm.simpleName
+import sift.core.element.AsmType
 
 data class TypeSignature(
     val type: ArgType,
@@ -18,10 +18,25 @@ data class TypeSignature(
 
         return "$type$array$inner"
     }
+
+    internal fun toTypeString(): String {
+        fun fromType(argType: ArgType): String = when (argType) {
+            is ArgType.Array -> fromType(argType.wrapped!!)
+            is ArgType.Plain -> argType.type.className
+            is ArgType.Var -> argType.type.name
+        }
+
+        val inner = args
+            .takeIf(MutableList<TypeSignature>::isNotEmpty)
+            ?.joinToString(prefix = "<", postfix = ">") { fromType(it.type) }
+            ?: ""
+
+        return "$type$inner"
+    }
 }
 
 sealed interface ArgType {
-    data class Plain(val type: Type) : ArgType {
+    data class Plain(val type: AsmType) : ArgType {
         override fun toString(): String = type.simpleName
     }